Refactor header construction WRT authorization
According to [1], "X-Emby-Authorization" and "X-MediaBrowser-Token"
headers are deprecated. Sadly, this is how this library authenticates
with a server. Switch to using the Authorization header, but also
remove some duplication with header construction so it is only done
in the HTTP code. Handle requests that do not include an AccessToken,
along with correcting how the connection manager creates its own API
object.

1: https://gist.github.com/nielsvanvelzen/ea047d9028f676185832e51ffaf12a6f